# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/services": {
      "get": {
        "operationId": "ServicesController_findActive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of active services"
          }
        },
        "summary": "Get all active services",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully"
          }
        },
        "summary": "Create a new service",
        "tags": [
          "services"
        ]
      }
    },
...
    "/api/v1/services/{id}": {
      "get": {
        "operationId": "ServicesController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details"
          }
        },
        "summary": "Get service by ID",
        "tags": [
          "services"
        ]
      },
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "Role of the caller",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully"
          }
        },
        "summary": "Update an existing service",
        "tags": [
          "services"
        ]
      }
    },
...
    "/api/v1/department-head/designations": {
      "get": {
        "operationId": "DepartmentHeadController_listDesignations",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-department-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List designations for a department",
        "tags": [
          "department-head"
        ]
      },
      "post": {
        "operationId": "DepartmentHeadController_createDesignation",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "DEPARTMENT_HEAD or STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDesignationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create department designation (roles, no levels)",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/designations/{id}": {
      "delete": {
        "operationId": "DepartmentHeadController_deleteDesignation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete designation",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/officers": {
      "get": {
        "operationId": "DepartmentHeadController_listOfficers",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-department-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List officers mapped to department",
        "tags": [
          "department-head"
        ]
      },
      "post": {
        "operationId": "DepartmentHeadController_onboardOfficer",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "DEPARTMENT_HEAD or STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardOfficerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Onboard officer mapped to exact jurisdiction node",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/officers/{id}/status": {
      "patch": {
        "operationId": "DepartmentHeadController_updateOfficerStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfficerStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Suspend / Unsuspend / Modify officer status",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/services": {
      "get": {
        "operationId": "DepartmentHeadController_listServices",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Browse department services",
        "tags": [
          "department-head"
        ]
      },
      "post": {
        "operationId": "DepartmentHeadController_createService",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "DEPARTMENT_HEAD",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDynamicServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create service with dynamic fields & workflow steps",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/services/{id}": {
      "get": {
        "operationId": "DepartmentHeadController_getService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get complete service specification & form configuration",
        "tags": [
          "department-head"
        ]
      },
      "put": {
        "operationId": "DepartmentHeadController_updateService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDynamicServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update existing dynamic service & workflow schema",
        "tags": [
          "department-head"
        ]
      },
      "delete": {
        "operationId": "DepartmentHeadController_deleteService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete dynamic service from department catalog",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/services/{id}/status": {
      "patch": {
        "operationId": "DepartmentHeadController_updateServiceStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Enable / Disable / Suspend service",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/analytics": {
      "get": {
        "operationId": "DepartmentHeadController_getDepartmentAnalytics",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-department-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Department operational analytics",
        "tags": [
          "department-head"
        ]
      }
    },
    "/api/v1/department-head/revenue": {
      "get": {
        "operationId": "DepartmentHeadController_getDepartmentRevenue",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-department-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Department financial revenue analytics & transaction ledger",
        "tags": [
          "department-head"
        ]
      }
    }
  },
  "info": {
    "title": "UCSDP Backend API",
    "description": "Unified Citizen Service and Delivery Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateServiceDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDesignationDto": {
        "type": "object",
        "properties": {}
      },
      "OnboardOfficerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOfficerStatusDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDynamicServiceDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}